Riddle works

+++ FrogDad/Scripts/DialogueBox.gd
@@ -21,7 +21,8 @@ onready var FrogDad = get_tree().get_root().find_node("FrogDad", true, false)
 
 var completed
 var dialogue_name
-
+var paused_for_riddle = false
+var current_riddle 
 	
 func _input(event):
 
@@ -29,19 +30,19 @@ func _input(event):
 	# the animation would be skipped and all text would show
 
 	if event.is_action_pressed("open") and content.visible:
+		if !paused_for_riddle:
+			if FrogDad:
+				MasterScript.frog_dad_state= "talking" # player can't move during dialogue
+			if text_animation.is_active():
 
-		if FrogDad:
-			MasterScript.frog_dad_state= "talking" # player can't move during dialogue
-		if text_animation.is_active():
-
-			text_animation.remove_all()
-			content.percent_visible = 1
-			_on_TextAnimation_tween_all_completed()
-		elif current + 1 < dialogs.size():
-
-			_show_dialog(current + 1)
-		else:
-			hide_dialog_box()
+				text_animation.remove_all()
+				content.percent_visible = 1
+				_on_TextAnimation_tween_all_completed()
+			elif current + 1 < dialogs.size():
+				
+					_show_dialog(current + 1)
+			else:
+				hide_dialog_box()
 
 
 func hide_dialog_box():
@@ -86,7 +87,7 @@ func show_new_talk_quest_notification_box():
 func show_new_quest_notifcation_box():
 	var HUDNode = FrogDad.find_node("HUD", true, false)
 	HUDNode.new_quest(MasterScript.currentQuestArray[0][1])
-	
+
 	
 func show_dialog_box(_dialogs, dialogueName):
 	$ColorRect.show()
@@ -125,7 +126,66 @@ func _show_dialog(index):
 	text_animation.start()
 	
 
-
-func _on_TextAnimation_tween_all_completed():
-	next_indicator.show()
 	
+
+	
+	
+func _on_TextAnimation_tween_all_completed():
+	if !MasterScript.odie_quest_active:
+		next_indicator.show()
+	else: 
+		$RiddleHUD.visible = true
+		paused_for_riddle = true
+
+# [{question, correct_answer : , wrong_answer1: , wrong_answer2: , question_name : ]
+func play_riddle (riddle):
+	current_riddle = riddle
+	show_dialog_box(riddle[0], riddle[1].question_name)
+	
+	var options = [riddle[1].correct_answer, riddle[1].wrong_answer1, riddle[1].wrong_answer2]
+	
+	var rng = RandomNumberGenerator.new()
+	
+	rng.randomize()
+	var rand_int = rng.randi_range(0, options.size()-1)#(randi() % options.size()) - 1
+	$RiddleHUD/VBoxContainer/Option1.set_text(options[rand_int])
+	options.remove(rand_int)
+	
+	rng.randomize()
+	rand_int = rng.randi_range(0, options.size()-1)
+	$RiddleHUD/VBoxContainer/Option2.set_text(options[rand_int])
+	options.remove(rand_int)
+	
+	rng.randomize()
+	rand_int = rng.randi_range(0, options.size()-1)
+	$RiddleHUD/VBoxContainer/Option3.set_text(options[rand_int])
+	options.remove(rand_int)
+	
+	
+	
+func _on_Option1_pressed():
+	_update_riddle($RiddleHUD/VBoxContainer/Option1)
+
+func _on_Option2_pressed():
+	_update_riddle($RiddleHUD/VBoxContainer/Option2)
+
+
+func _on_Option3_pressed():
+	_update_riddle($RiddleHUD/VBoxContainer/Option3)
+
+func _update_riddle(button):
+	$RiddleHUD.visible = false
+	paused_for_riddle = false
+	if button.text == current_riddle[1].correct_answer:
+		if button.text == "Promise":
+			play_riddle([[{avatar = "gertrude", text = "What flies when it's born, lies when it's alive, and runs when it's dead?" }],
+			{correct_answer = "A snowflake", wrong_answer1 = "A fly?", wrong_answer2 = "Lies when it's alive? My ex wife", question_name = "keyword"}])
+		if button.text == "A snowflake":
+			play_riddle([[{avatar = "gertrude", text = "What is always coming but never arrives?" }],
+			{correct_answer = "Tomorrow", wrong_answer1 = "My next paycheck!", wrong_answer2 = "My SnailMail food delivery", question_name = "keyword"}])
+	else:
+		MasterScript.odie_quest_active = false
+		if button.text == "A fly?":
+			show_dialog_box([{avatar = "gertrude", text = "Nice try. I did know a fly who liked to lie. He was a conniving son of a birch! Did I ever tell you the story about-"},
+			{avatar = "frogDad", text = "I'm gonna stop you right there, pal. I think I left my stove on, catch you later!"},
+			{avatar = "gertrude", text = "What? Okay. Well, let me know if you wanna try again!"}], "IDJOWIJ")
